+++ src/mesh/mesh.hpp
@@ -153,8 +153,9 @@ class Mesh {
   void ApplyUserWorkBeforeOutput(ParameterInput *pin);
 
   // Boundary Functions
-  BValFunc MeshBndryFnctn[6];
-  SBValFunc SwarmBndryFnctn[6];
+  BValFunc MeshBndryFnctn[BOUNDARY_NFACES];
+  SBValFunc SwarmBndryFnctn[BOUNDARY_NFACES];
+  std::array<std::vector<BValFunc>, BOUNDARY_NFACES> UserBoundaryFunctions;
 
   // defined in either the prob file or default_pgen.cpp in ../pgen/
   std::function<void(Mesh *, ParameterInput *, MeshData<Real> *)> ProblemGenerator =